refactor: move search components into search folder
