--- frontend/src/components/SearchResults/index.ts
@@ -1,4 +1,0 @@
-export { default as SearchLoadingState } from './SearchLoadingState';
-export { default as SearchErrorState } from './SearchErrorState';
-export { default as SongsView } from './SongsView';
-export { default as SearchResultsStateHandler } from './SearchResultsStateHandler';
