+++ frontend/src/search/components/SearchResults/SearchResults.types.ts
