+++ frontend/src/search/components/SearchResults/SearchResultsLayout/SearchResultsLayout.types.ts
@@ -1,0 +1,13 @@
+/**
+ * Props interface for SearchResultsLayout component
+ */
+import type { Artist, SearchType, Song } from "@chordium/types";
+
+export interface SearchResultsLayoutProps {
+  results: (Artist | Song)[];
+  searchType?: SearchType;
+  onView: (song: Song) => void;
+  onDelete: (songId: string) => void;
+  onArtistSelect?: (artist: Artist) => void;
+  hasSearched?: boolean;
+}
