+++ frontend/src/search/components/SearchResults/SearchResultsSection/SearchResultsSection.types.ts
@@ -1,0 +1,9 @@
+/**
+ * Props interface for SearchResultsSection component
+ */
+
+export interface SearchResultsSectionProps {
+  title: string;
+  children: React.ReactNode;
+  className?: string;
+}
