+++ frontend/src/search/components/SearchResults/SongsView.tsx
@@ -1,11 +1,10 @@
 import React, { useCallback } from 'react';
-import { Song } from '@/types/song';
-import { Artist } from '@/types/artist';
+import { Artist, Song } from '@chordium/types';
 import ResultCard from '@/components/ResultCard';
 import VirtualizedListWithArrow from '@/components/ui/VirtualizedListWithArrow';
 import { ListChildComponentProps } from 'react-window';
 import { CARD_HEIGHTS } from '@/constants/ui-constants';
-import SearchResultsSection from '../SearchResultsSection';
+import { SearchResultsSection } from '.';
 import { cyAttr } from '@/utils/test-utils';
 import '@/components/custom-scrollbar.css';
