+++ frontend/src/search/components/SearchResults/index.ts
@@ -1,0 +1,5 @@
+export { default as SearchResultsLayout } from "./SearchResultsLayout/SearchResultsLayout";
+export { default as SearchResultsSection } from "./SearchResultsSection/SearchResultsSection";
+export { default as SearchResultsStateHandler } from "./SearchResultsStateHandler/SearchResultsStateHandler";
+export { default as SearchResults } from "./SearchResults";
+export { default as SongsView } from "./SongsView";
