+++ frontend/src/search/components/index.ts
@@ -1,2 +1,2 @@
+export * from "./SearchResults";
 export * from "./SearchTab";
-export { default as SearchResults } from "./SearchResults";
